Fix notification groups
* Fixed group notification left when click on last action in group
* Fixed internal group prefix "group_" is shown on notification
